[chore] refactor stdout publisher
Move stdout publisher to its own package. Publish function returns an error
