+++ pkg/metrics/publisher/expvar/expvar.go
@@ -74,12 +74,14 @@ func (exp *Expvar) Stop(shutdownTimeout time.Duration) {
 }
 
 // Publish is called by the publisher goroutine and saves the raw stats.
-func (exp *Expvar) Publish(data map[string]any) {
+func (exp *Expvar) Publish(data map[string]any) error {
 	exp.mu.Lock()
 	{
 		exp.data = maps.Clone(data)
 	}
 	exp.mu.Unlock()
+
+	return nil
 }
 
 // handler is what consumers call to get the raw stats.
